functions/update_weather_data/requirements.txt
requests
google-cloud-firestore